Refine backend AIP

+++ backend/controllers/product.controller.js
@@ -52,6 +52,10 @@ export const deleteProduct = async(req, res) => {
   
     const { id } = req.params;
     console.log(id);
+
+    if (!mongoose.Types.ObjectId.isValid(id)) {
+      return res.status(404).send('Invalid product id');
+    }
     
     try {
       await Product.findByIdAndDelete(id);
@@ -59,7 +63,7 @@ export const deleteProduct = async(req, res) => {
 
     } catch (error) {
       console.error("Error in deleting product:", error.message);
-      res.status(404).json({ success: false, message: "Product not found" });
+      res.status(500).json({ success: false, message: "Server Error" });
     }
 
 }
